Expose test_utils from sqlite store via a feature flag
