+++ p2panda-store/Cargo.toml
@@ -19,13 +19,14 @@ workspace = true
 default = ["memory"]
 memory = []
 sqlite = ["dep:ciborium", "dep:sqlx", "dep:hex"]
+test_utils = ["dep:rand"]
 
 [dependencies]
 ciborium = { version = "0.2.2", optional = true }
 hex = { version = "0.4.3", optional = true }
 p2panda-core = { path = "../p2panda-core", version = "0.2.0" }
-sqlx = { version = "0.8.3", optional = true, features = ["sqlite",
-"runtime-tokio"] }
+rand = { version = "0.8.5", optional = true }
+sqlx = { version = "0.8.3", optional = true, features = ["sqlite", "runtime-tokio"] }
 thiserror = "2.0.11"
 trait-variant = "0.1.2"
