+++ p2panda-store/src/sqlite/mod.rs
@@ -2,3 +2,5 @@
 
 mod models;
 pub mod store;
+#[cfg(any(test, feature = "test_utils"))]
+pub mod test_utils;
